Moved files into structure
