+++ biome.json
@@ -60,7 +60,8 @@
         "noHeadElement": "off",
         "noExportedImports": "off",
         "noDocumentCookie": "off",
-        "noProcessEnv": "off"
+        "noProcessEnv": "off",
+        "useComponentExportOnlyModules": "off"
       },
       "performance": {
         "all": true,
